+++ docs/changelog.md
@@ -13,7 +13,9 @@ and this project adheres to
 
 - `meeple import` - Import BoardGameGeek user collections.
   - Import as either one large collection or as separate collections based on item status.
-  - `--dry-run` - Summarize potential import operations without persisting.
+  - `--one` - Import as one collection.
+  - `--many` - Import as separate collections by status.
+  - `--dry-run` - Simulate import operations without persisting.
   - `-v/--verbose` - Output additional details.
 
 ## [v0.1.0] - 2023-05-16